Move StoreDatabase into database folder, merge with existing changes
