+++ app/src/main/java/com/example/b07group7project/Order.java
@@ -1,10 +1,10 @@
 package com.example.b07group7project;
 
+import com.example.b07group7project.database_abstractions.Store;
+
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import com.example.b07group7project.database_abstractions.Store;
 
 public class Order {
     HashMap<Product, Integer> productsuuidToQuantity;
@@ -37,5 +37,13 @@ public class Order {
             newhashmap.put(m.getKey().uuid, m.getValue());
         }
         return newhashmap;
+    }
+
+    public Store getStore(){
+        return store;
+    }
+
+    public String getUuid() {
+        return uuid;
     }
 }
